+++ CHANGELOG.md
@@ -7,5 +7,6 @@
 - Add RoboCop::Cop::Vicenzo::RSpec::NestedSubjectRedefinition #2;
 - Add RuboCop::Cop::Vicenzo::Rails::EnumInclusionOfValidation #3;
 - Add RuboCop::Cop::Vicenzo::RSpec::NestedContextImproperStart #4;
+- Add RuboCop::Cop::Vicenzo::RSpec::MixedExampleGroups #6;
 
 - Change RuboCop::Cop::Vicenzo::RSpec::NestedContextImproperStart inherits from Rspec::Base #5;